@@ -15,7 +15,7 @@
 
 [tool.poetry]
 name = "neo4j-graphrag"
-version = "0.6.3"
+version = "0.7.0"
 description = "Python package to allow easy integration to Neo4j's GraphRAG features"
 authors = ["Neo4j, Inc <team-gen-ai@neo4j.com>"]
 license = "Apache License, Version 2.0"
@@ -28,7 +28,7 @@ include = "neo4j_graphrag"
 from = "src"
 
 [tool.poetry.dependencies]
-python = "^3.8.1"
+python = "^3.9.0"
 neo4j = "^5.17.0"
 pydantic = "^2.6.3"
 weaviate-client = {version = "^4.6.1", optional = true}
