@@ -33,12 +33,14 @@ from neo4j_graphrag.experimental.components.text_splitters.fixed_size_splitter i
     FixedSizeSplitter,
 )
 from neo4j_graphrag.experimental.pipeline.pipeline import PipelineResult
-from neo4j_graphrag.llm import OpenAILLM
+from neo4j_graphrag.llm import LLMInterface, OpenAILLM
 
 logging.basicConfig(level=logging.INFO)
 
 
-async def main(neo4j_driver: neo4j.Driver) -> PipelineResult:
+async def define_and_run_pipeline(
+    neo4j_driver: neo4j.AsyncDriver, llm: LLMInterface
+) -> PipelineResult:
     from neo4j_graphrag.experimental.pipeline import Pipeline
 
     # Instantiate Entity and Relation objects
@@ -86,13 +88,7 @@ async def main(neo4j_driver: neo4j.Driver) -> PipelineResult:
     pipe.add_component(SchemaBuilder(), "schema")
     pipe.add_component(
         LLMEntityRelationExtractor(
-            llm=OpenAILLM(
-                model_name="gpt-4o",
-                model_params={
-                    "max_tokens": 2000,
-                    "response_format": {"type": "json_object"},
-                },
-            ),
+            llm=llm,
             on_error=OnError.RAISE,
         ),
         "extractor",
@@ -127,8 +123,23 @@ async def main(neo4j_driver: neo4j.Driver) -> PipelineResult:
     return await pipe.run(pipe_inputs)
 
 
-if __name__ == "__main__":
-    with neo4j.GraphDatabase.driver(
+async def main() -> PipelineResult:
+    llm = OpenAILLM(
+        model_name="gpt-4o",
+        model_params={
+            "max_tokens": 2000,
+            "response_format": {"type": "json_object"},
+        },
+    )
+    driver = neo4j.AsyncGraphDatabase.driver(
         "bolt://localhost:7687", auth=("neo4j", "password")
-    ) as driver:
-        print(asyncio.run(main(driver)))
+    )
+    res = await define_and_run_pipeline(driver, llm)
+    await driver.close()
+    await llm.async_client.close()
+    return res
+
+
+if __name__ == "__main__":
+    res = asyncio.run(main())
+    print(res)
