@@ -28,6 +28,9 @@ from neo4j_graphrag.experimental.components.entity_relation_extractor import (
     OnError,
 )
 from neo4j_graphrag.experimental.components.kg_writer import Neo4jWriter
+from neo4j_graphrag.experimental.components.resolver import (
+    SinglePropertyExactMatchResolver,
+)
 from neo4j_graphrag.experimental.components.schema import (
     SchemaBuilder,
     SchemaEntity,
@@ -63,7 +66,7 @@ def schema_builder() -> SchemaBuilder:
 
 @pytest.fixture
 def text_splitter() -> FixedSizeSplitter:
-    return FixedSizeSplitter(chunk_size=500, chunk_overlap=100)
+    return FixedSizeSplitter(chunk_size=500, chunk_overlap=10)
 
 
 @pytest.fixture
@@ -85,35 +88,27 @@ def kg_writer(driver: neo4j.Driver) -> Neo4jWriter:
 
 
 @pytest.fixture
+def entity_resolver(driver: neo4j.Driver) -> SinglePropertyExactMatchResolver:
+    return SinglePropertyExactMatchResolver(driver)
+
+
+@pytest.fixture
 def kg_builder_pipeline(
     text_splitter: FixedSizeSplitter,
     chunk_embedder: TextChunkEmbedder,
     schema_builder: SchemaBuilder,
     entity_relation_extractor: LLMEntityRelationExtractor,
     kg_writer: Neo4jWriter,
+    entity_resolver: SinglePropertyExactMatchResolver,
 ) -> Pipeline:
     pipe = Pipeline()
     # define the components
-    pipe.add_component(
-        text_splitter,
-        "splitter",
-    )
-    pipe.add_component(
-        chunk_embedder,
-        "embedder",
-    )
-    pipe.add_component(
-        schema_builder,
-        "schema",
-    )
-    pipe.add_component(
-        entity_relation_extractor,
-        "extractor",
-    )
-    pipe.add_component(
-        kg_writer,
-        "writer",
-    )
+    pipe.add_component(text_splitter, "splitter")
+    pipe.add_component(chunk_embedder, "embedder")
+    pipe.add_component(schema_builder, "schema")
+    pipe.add_component(entity_relation_extractor, "extractor")
+    pipe.add_component(kg_writer, "writer")
+    pipe.add_component(entity_resolver, "resolver")
     # define the execution order of component
     # and how the output of previous components must be used
     pipe.connect("splitter", "embedder", input_config={"text_chunks": "splitter"})
@@ -125,12 +120,31 @@ def kg_builder_pipeline(
         "writer",
         input_config={"graph": "extractor"},
     )
+    pipe.connect("writer", "resolver", {})
     return pipe
 
 
 @pytest.fixture
 def harry_potter_text() -> str:
     with open(os.path.join(BASE_DIR, "data/harry_potter.txt"), "r") as f:
+        text = f.read()
+    return text
+
+
+@pytest.fixture
+def harry_potter_text_part1() -> str:
+    with open(
+        os.path.join(BASE_DIR, "data/documents/harry_potter_part1.txt"), "r"
+    ) as f:
+        text = f.read()
+    return text
+
+
+@pytest.fixture
+def harry_potter_text_part2() -> str:
+    with open(
+        os.path.join(BASE_DIR, "data/documents/harry_potter_part2.txt"), "r"
+    ) as f:
         text = f.read()
     return text
 
@@ -189,7 +203,6 @@ async def test_pipeline_builder_happy_path(
                         ]
                     }"""
         ),
-        LLMResponse(content='{"nodes": [], "relationships": []}'),
         LLMResponse(content='{"nodes": [], "relationships": []}'),
     ]
 
@@ -252,25 +265,25 @@ async def test_pipeline_builder_happy_path(
     }
     res = await kg_builder_pipeline.run(pipe_inputs)
     # llm must have been called for each chunk
-    assert llm.ainvoke.call_count == 3
+    assert llm.ainvoke.call_count == 2
     # result must be success
     assert isinstance(res, PipelineResult)
     assert res.run_id is not None
-    assert res.result == {"writer": {"status": "SUCCESS"}}
+    assert "resolver" in res.result
     # check component's results
     chunks = await kg_builder_pipeline.store.get_result_for_component(
         res.run_id, "splitter"
     )
-    assert len(chunks["chunks"]) == 3
+    assert len(chunks["chunks"]) == 2
     graph = await kg_builder_pipeline.store.get_result_for_component(
         res.run_id, "extractor"
     )
-    # 3 entities + 3 chunks + 1 document
+    # 3 entities + 2 chunks + 1 document
     nodes = graph["nodes"]
-    assert len(nodes) == 7
+    assert len(nodes) == 6
     label_counts = dict(Counter([n["label"] for n in nodes]))
     assert label_counts == {
-        "Chunk": 3,
+        "Chunk": 2,
         "Document": 1,
         "Person": 2,
         "Organization": 1,
@@ -279,23 +292,23 @@ async def test_pipeline_builder_happy_path(
     # + 3 rels between entities and their chunk
     # + 2 "NEXT_CHUNK" rels
     relationships = graph["relationships"]
-    assert len(relationships) == 10
+    assert len(relationships) == 8
     type_counts = dict(Counter([r["type"] for r in relationships]))
     assert type_counts == {
         "FROM_CHUNK": 3,
-        "FROM_DOCUMENT": 3,
+        "FROM_DOCUMENT": 2,
         "KNOWS": 1,
         "LED_BY": 1,
-        "NEXT_CHUNK": 2,
+        "NEXT_CHUNK": 1,
     }
     # then check content of neo4j db
     created_nodes = driver.execute_query("MATCH (n) RETURN n")
-    assert len(created_nodes.records) == 7
+    assert len(created_nodes.records) == 6
     created_rels = driver.execute_query("MATCH ()-[r]->() RETURN r")
-    assert len(created_rels.records) == 10
+    assert len(created_rels.records) == 8
 
     created_chunks = driver.execute_query("MATCH (n:Chunk) RETURN n").records
-    assert len(created_chunks) == 3
+    assert len(created_chunks) == 2
     for c in created_chunks:
         node = c.get("n")
         assert node.get("embedding") == [1, 2, 3]
@@ -358,7 +371,6 @@ async def test_pipeline_builder_failing_chunk_raise(
                     }"""
         ),
         LLMResponse(content="invalid json"),
-        LLMResponse(content='{"nodes": [], "relationships": []}'),
     ]
 
     # user input:
@@ -437,7 +449,6 @@ async def test_pipeline_builder_failing_chunk_do_not_raise(
                         ]
                     }"""
         ),
-        LLMResponse(content='{"nodes": [], "relationships": []}'),
     ]
 
     # user input:
@@ -458,37 +469,157 @@ async def test_pipeline_builder_failing_chunk_do_not_raise(
     ).component.on_error = OnError.IGNORE  # type: ignore[attr-defined, unused-ignore]
     res = await kg_builder_pipeline.run(pipe_inputs)
     # llm must have been called for each chunk
-    assert llm.ainvoke.call_count == 3
+    assert llm.ainvoke.call_count == 2
     # result must be success
     assert isinstance(res, PipelineResult)
     assert res.run_id is not None
-    assert res.result == {"writer": {"status": "SUCCESS"}}
+    assert res.result == {
+        "resolver": {"number_of_created_nodes": 3, "number_of_nodes_to_resolve": 3}
+    }
     # check component's results
     chunks = await kg_builder_pipeline.store.get_result_for_component(
         res.run_id, "splitter"
     )
-    assert len(chunks["chunks"]) == 3
+    assert len(chunks["chunks"]) == 2
     graph = await kg_builder_pipeline.store.get_result_for_component(
         res.run_id, "extractor"
     )
-    # 3 entities + 3 chunks
+    # 3 entities + 2 chunks
     nodes = graph["nodes"]
-    assert len(nodes) == 6
+    assert len(nodes) == 5
     label_counts = dict(Counter([n["label"] for n in nodes]))
     assert label_counts == {
-        "Chunk": 3,
+        "Chunk": 2,
         "Person": 2,
         "Organization": 1,
     }
     # 2 relationships between entities
     # + 3 rels between entities and their chunk
-    # + 2 "NEXT_CHUNK" rels
+    # + 1 "NEXT_CHUNK" rels
     relationships = graph["relationships"]
-    assert len(relationships) == 7
+    assert len(relationships) == 6
     type_counts = dict(Counter([r["type"] for r in relationships]))
-    assert type_counts == {"FROM_CHUNK": 3, "KNOWS": 1, "LED_BY": 1, "NEXT_CHUNK": 2}
+    assert type_counts == {"FROM_CHUNK": 3, "KNOWS": 1, "LED_BY": 1, "NEXT_CHUNK": 1}
     # then check content of neo4j db
     created_nodes = driver.execute_query("MATCH (n) RETURN n")
-    assert len(created_nodes.records) == 6
+    assert len(created_nodes.records) == 5
     created_rels = driver.execute_query("MATCH ()-[r]->() RETURN r")
-    assert len(created_rels.records) == 7
+    assert len(created_rels.records) == 6
+
+
+@pytest.mark.asyncio
+@pytest.mark.usefixtures("setup_neo4j_for_kg_construction")
+async def test_pipeline_builder_two_documents(
+    harry_potter_text_part1: str,
+    harry_potter_text_part2: str,
+    embedder: MagicMock,
+    llm: MagicMock,
+    driver: neo4j.Driver,
+    kg_builder_pipeline: Pipeline,
+) -> None:
+    """Run same pipeline on two documents. Check entity resolution.
+
+    First document:
+    2 chunks, entities Harry and The Order of the Phoenix, 1 relationship
+    Second document:
+    1 chunk, entities Harry and Alastor Mad-Eye Moody, 1 relationship
+
+    Should create:
+    1 document node
+    3 chunk nodes
+    3 entities (1 Harry + the other two)
+    ==> 7 nodes
+    3 relationships for lexical graph + 3 relationships for the entity graph
+    ==> 6 relationships
+    """
+    driver.execute_query("MATCH (n) DETACH DELETE n")
+    embedder.embed_query.return_value = [1, 2, 3]
+    llm.ainvoke.side_effect = [
+        LLMResponse(
+            content="""{
+                        "nodes": [
+                            {
+                                "id": "0",
+                                "label": "Person",
+                                "properties": {
+                                    "name": "Harry"
+                                }
+                            },
+                            {
+                                "id": "1",
+                                "label": "Organization",
+                                "properties": {
+                                    "name": "The Order of the Phoenix"
+                                }
+                            }
+                        ],
+                        "relationships": [
+                            {
+                                "type": "MEMBER_OF",
+                                "start_node_id": "0",
+                                "end_node_id": "1"
+                            }
+                        ]
+                    }"""
+        ),
+        LLMResponse(
+            content="""{
+                        "nodes": [
+                            {
+                                "id": "10",
+                                "label": "Person",
+                                "properties": {
+                                    "name": "Harry"
+                                }
+                            },
+                            {
+                                "id": "11",
+                                "label": "Person",
+                                "properties": {
+                                    "name": "Alastor Mad-Eye Moody"
+                                }
+                            }
+                        ],
+                        "relationships": [
+                            {
+                                "type": "KNOWS",
+                                "start_node_id": "10",
+                                "end_node_id": "11"
+                            }
+                        ]
+                    }"""
+        ),
+        LLMResponse(content='{"nodes": [], "relationships": []}'),
+    ]
+
+    # user input:
+    # the initial text
+    # and the list of entities and relations we are looking for
+    pipe_inputs_1 = {
+        "splitter": {"text": harry_potter_text_part1},
+        # note: schema not used in this test because
+        # we are mocking the LLM
+        "schema": {
+            "entities": [],
+            "relations": [],
+            "potential_schema": [],
+        },
+    }
+    pipe_inputs_2 = {
+        "splitter": {"text": harry_potter_text_part2},
+        # note: schema not used in this test because
+        # we are mocking the LLM
+        "schema": {
+            "entities": [],
+            "relations": [],
+            "potential_schema": [],
+        },
+    }
+    await kg_builder_pipeline.run(pipe_inputs_1)
+    await kg_builder_pipeline.run(pipe_inputs_2)
+    created_nodes = driver.execute_query("MATCH (n:__Entity__) RETURN n")
+    assert len(created_nodes.records) == 3
+    created_rels = driver.execute_query(
+        "MATCH (:__Entity__)-[r]->(:__Entity__) RETURN r"
+    )
+    assert len(created_rels.records) == 2
