@@ -28,26 +28,39 @@ from neo4j_graphrag.experimental.components.types import (
 async def test_kg_writer(driver: neo4j.Driver) -> None:
     start_node = Neo4jNode(
         id="1",
-        label="Document",
+        label="MyLabel",
         properties={"chunk": 1},
         embedding_properties={"vectorProperty": [1.0, 2.0, 3.0]},
     )
     end_node = Neo4jNode(
         id="2",
-        label="Document",
-        properties={"chunk": 2},
-        embedding_properties={"vectorProperty": [1.0, 2.0, 3.0]},
+        label="MyLabel",
+        properties={},
+        embedding_properties=None,
     )
     relationship = Neo4jRelationship(
-        start_node_id="1", end_node_id="2", type="NEXT_CHUNK"
+        start_node_id="1", end_node_id="2", type="MY_RELATIONSHIP"
     )
-    graph = Neo4jGraph(nodes=[start_node, end_node], relationships=[relationship])
+    node_with_two_embeddings = Neo4jNode(
+        id="3",
+        label="MyLabel",
+        properties={"chunk": 1},
+        embedding_properties={
+            "vectorProperty": [1.0, 2.0, 3.0],
+            "otherVectorProperty": [10.0, 20.0, 30.0],
+        },
+    )
+    graph = Neo4jGraph(
+        nodes=[start_node, end_node, node_with_two_embeddings],
+        relationships=[relationship],
+    )
 
     neo4j_writer = Neo4jWriter(driver=driver)
-    await neo4j_writer.run(graph=graph)
+    res = await neo4j_writer.run(graph=graph)
+    assert res.status == "SUCCESS"
 
     query = """
-    MATCH (a:Document {id: '1'})-[r:NEXT_CHUNK]->(b:Document {id: '2'})
+    MATCH (a:MyLabel {id: '1'})-[r:MY_RELATIONSHIP]->(b:MyLabel {id: '2'})
     RETURN a, r, b
     """
     record = driver.execute_query(query).records[0]
@@ -56,11 +69,10 @@ async def test_kg_writer(driver: neo4j.Driver) -> None:
     node_a = record["a"]
     assert start_node.label in list(node_a.labels)
     assert start_node.id == str(node_a.get("id"))
-    if start_node.properties:
-        for key, val in start_node.properties.items():
-            assert key in node_a.keys()
-            assert val == node_a.get(key)
-    if start_node.embedding_properties:
+    for key, val in start_node.properties.items():
+        assert key in node_a.keys()
+        assert val == node_a.get(key)
+    if start_node.embedding_properties:  # for mypy
         for key, val in start_node.embedding_properties.items():
             assert key in node_a.keys()
             assert node_a.get(key) == [1.0, 2.0, 3.0]
@@ -68,17 +80,23 @@ async def test_kg_writer(driver: neo4j.Driver) -> None:
     node_b = record["b"]
     assert end_node.label in list(node_b.labels)
     assert end_node.id == str(node_b.get("id"))
-    if end_node.properties:
-        for key, val in end_node.properties.items():
-            assert key in node_b.keys()
-            assert val == node_b.get(key)
-    if end_node.embedding_properties:
-        for key, val in end_node.embedding_properties.items():
-            assert key in node_b.keys()
-            assert node_b.get(key) == [1.0, 2.0, 3.0]
+    for key, val in end_node.properties.items():
+        assert key in node_b.keys()
+        assert val == node_b.get(key)
 
     rel = record["r"]
     assert rel.type == relationship.type
-    assert relationship.start_node_id and relationship.end_node_id in [
-        str(node.get("id")) for node in rel.nodes
-    ]
+    assert relationship.start_node_id == rel.start_node.get("id")
+    assert relationship.end_node_id == rel.end_node.get("id")
+
+    query = """
+    MATCH (c:MyLabel {id: '3'})
+    RETURN c
+    """
+    records = driver.execute_query(query).records
+    assert len(records) == 1
+    node_c = records[0]["c"]
+    if node_with_two_embeddings.embedding_properties:  # for mypy
+        for key, val in node_with_two_embeddings.embedding_properties.items():
+            assert key in node_c.keys()
+            assert val == node_c.get(key)
